# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
  models.AssignPermissionRequest:
    properties:
      permissionID:
        type: string
      relationType:
        type: string
    type: object
...
  models.CheckPermissionRequest:
    properties:
      object_id:
        type: string
      permission_name:
        type: string
      subject_id:
        type: string
    required:
    - object_id
    - permission_name
    - subject_id
    type: object
  models.CheckPermissionResponse:
    properties:
      has_permission:
        type: boolean
    type: object
...
  models.CreateRelationRequest:
    properties:
      relationType:
        type: string
      sourceID:
        type: string
      targetID:
        type: string
    type: object
...
  models.DeleteRelationRequest:
    properties:
      relationType:
        type: string
      sourceID:
        type: string
      targetID:
        type: string
    type: object
...
  models.Entity:
    properties:
      createdAt:
        type: string
      id:
        type: string
      type:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Permission:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.Relation:
    properties:
      createdAt:
        type: string
      id:
        type: string
      relationType:
        type: string
      sourceID:
        type: string
      targetID:
        type: string
    type: object
  models.RevokePermissionRequest:
    properties:
      permissionID:
        type: string
      relationType:
        type: string
    type: object
...
  models.UserRoleResponse:
    properties:
      role:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: API для платежной системы
  title: Payment API
  version: "1.0"
paths:
...
  /role/entities:
    get:
      description: Возвращает список всех сущностей (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Entity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить все сущности
      tags:
      - role
  /role/permissions:
    get:
      description: Возвращает список всех доступных разрешений (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить все разрешения
      tags:
      - role
  /role/permissions/assign:
    post:
      consumes:
      - application/json
      description: Назначает разрешение для указанного типа связи (Admin only)
      parameters:
      - description: Данные разрешения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AssignPermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Назначить разрешение
      tags:
      - role
  /role/permissions/check:
    post:
      consumes:
      - application/json
      description: Проверяет наличие указанного разрешения у субъекта для объекта
        (Admin only)
      parameters:
      - description: Данные проверки разрешения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CheckPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CheckPermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Проверить разрешение
      tags:
      - role
  /role/permissions/revoke:
    delete:
      consumes:
      - application/json
      description: Отзывает разрешение для указанного типа связи (Admin only)
      parameters:
      - description: Данные разрешения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RevokePermissionRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Отозвать разрешение
      tags:
      - role
  /role/relations:
    delete:
      consumes:
      - application/json
      description: Удаляет связь между source и target сущностями (Admin only)
      parameters:
      - description: Данные связи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteRelationRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Удалить связь
      tags:
      - role
    post:
      consumes:
      - application/json
      description: Создает связь между source и target сущностями (Admin only)
      parameters:
      - description: Данные связи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateRelationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Создать связь
      tags:
      - role
  /role/relations/{relationType}/permissions:
    get:
      description: Возвращает список разрешений для указанного типа связи (Admin only)
      parameters:
      - description: Тип связи
        in: path
        name: relationType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить разрешения для типа связи
      tags:
      - role
  /role/role:
    get:
      description: Возвращает роль текущего аутентифицированного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRoleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить роль пользователя
      tags:
      - role
  /role/users/{userID}/permissions:
    get:
      description: Возвращает список всех разрешений указанного пользователя (Admin
        only)
      parameters:
      - description: ID пользователя
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить разрешения пользователя
      tags:
      - role
  /role/users/{userID}/relations:
    get:
      description: Возвращает список всех связей указанного пользователя (Admin only)
      parameters:
      - description: ID пользователя
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Relation'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить связи пользователя
      tags:
      - role
...
schemes:
- https
securityDefinitions:
  CookieAuth:
    description: Аутентификация через куки (токен автоматически отправляется сервером)
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"